requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
transformers==4.44.0
torch==2.4.0
sentence-transformers==3.0.1
chromadb==0.5.5
pypdf2==3.0.1
python-docx==1.1.0
scikit-learn==1.5.1
numpy==1.26.4
pydantic==2.9.0
jinja2==3.1.4
aiofiles==24.1.0
python-jose==3.3.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
